refactiry transform method
