+++ src/route-handlers/create-schedule/helpers/transform-create-schedule-body-to-grpc-input.ts
@@ -1,11 +1,13 @@
+import isEmpty from 'lodash/isEmpty';
+
 import { type CreateScheduleRequest__Input } from '@/__generated__/proto-ts/uber/cadence/api/v1/CreateScheduleRequest';
+import { type RetryPolicy__Input } from '@/__generated__/proto-ts/uber/cadence/api/v1/RetryPolicy';
 import getGrpcTimestampFromIso from '@/utils/datetime/get-grpc-timestamp-from-iso';
 
 import processWorkflowInput from '../../start-workflow/helpers/process-workflow-input';
 import { DEFAULT_TASK_START_TO_CLOSE_TIMEOUT_SECONDS } from '../../start-workflow/start-workflow.constants';
 import { type CreateScheduleRequestBody } from '../create-schedule.types';
 
-import mapRetryPolicyToGrpcInput from './map-retry-policy-to-grpc-input';
 import secondsToGrpcDurationInput from './seconds-to-grpc-duration-input';
 
 export default function transformCreateScheduleBodyToGrpcInput({
@@ -42,59 +44,77 @@ export default function transformCreateScheduleBodyToGrpcInput({
     workerSDKLanguage: sw.workerSDKLanguage,
   });
 
+  const retryPolicy = sw.retryPolicy;
+  const grpcRetryPolicy: RetryPolicy__Input | undefined =
+    !retryPolicy || isEmpty(retryPolicy)
+      ? undefined
+      : {
+          initialInterval: retryPolicy.initialIntervalSeconds
+            ? secondsToGrpcDurationInput(retryPolicy.initialIntervalSeconds)
+            : undefined,
+          backoffCoefficient: retryPolicy.backoffCoefficient,
+          maximumInterval: retryPolicy.maximumIntervalSeconds
+            ? secondsToGrpcDurationInput(retryPolicy.maximumIntervalSeconds)
+            : undefined,
+          expirationInterval: retryPolicy.expirationIntervalSeconds
+            ? secondsToGrpcDurationInput(retryPolicy.expirationIntervalSeconds)
+            : undefined,
+          maximumAttempts: retryPolicy.maximumAttempts,
+        };
+
+  const spec = {
+    cronExpression,
+    startTime: startTime ? getGrpcTimestampFromIso(startTime) : undefined,
+    endTime: endTime ? getGrpcTimestampFromIso(endTime) : undefined,
+    jitter: jitterSeconds
+      ? secondsToGrpcDurationInput(jitterSeconds)
+      : undefined,
+  };
+
+  const startWorkflow = {
+    workflowType: sw.workflowType,
+    taskList: sw.taskList,
+    input: processedInput
+      ? { data: Buffer.from(processedInput, 'utf-8') }
+      : undefined,
+    workflowIdPrefix: sw.workflowIdPrefix,
+    executionStartToCloseTimeout: secondsToGrpcDurationInput(
+      sw.executionStartToCloseTimeoutSeconds
+    ),
+    taskStartToCloseTimeout: secondsToGrpcDurationInput(
+      sw.taskStartToCloseTimeoutSeconds ??
+        DEFAULT_TASK_START_TO_CLOSE_TIMEOUT_SECONDS
+    ),
+    retryPolicy: grpcRetryPolicy,
+    memo: sw.memo
+      ? {
+          fields: Object.fromEntries(
+            Object.entries(sw.memo).map(([k, v]) => [
+              k,
+              { data: Buffer.from(JSON.stringify(v), 'utf-8') },
+            ])
+          ),
+        }
+      : undefined,
+    searchAttributes: sw.searchAttributes
+      ? {
+          indexedFields: Object.fromEntries(
+            Object.entries(sw.searchAttributes).map(([k, v]) => [
+              k,
+              { data: Buffer.from(JSON.stringify(v), 'utf-8') },
+            ])
+          ),
+        }
+      : undefined,
+  };
+
   return {
     domain,
     scheduleId,
 
-    // Schedule spec
-    spec: {
-      cronExpression,
-      startTime: startTime ? getGrpcTimestampFromIso(startTime) : undefined,
-      endTime: endTime ? getGrpcTimestampFromIso(endTime) : undefined,
-      jitter: jitterSeconds
-        ? secondsToGrpcDurationInput(jitterSeconds)
-        : undefined,
-    },
+    spec,
 
-    // Start-workflow action
-    action: {
-      startWorkflow: {
-        workflowType: sw.workflowType,
-        taskList: sw.taskList,
-        input: processedInput
-          ? { data: Buffer.from(processedInput, 'utf-8') }
-          : undefined,
-        workflowIdPrefix: sw.workflowIdPrefix,
-        executionStartToCloseTimeout: secondsToGrpcDurationInput(
-          sw.executionStartToCloseTimeoutSeconds
-        ),
-        taskStartToCloseTimeout: secondsToGrpcDurationInput(
-          sw.taskStartToCloseTimeoutSeconds ??
-            DEFAULT_TASK_START_TO_CLOSE_TIMEOUT_SECONDS
-        ),
-        retryPolicy: mapRetryPolicyToGrpcInput(sw.retryPolicy),
-        memo: sw.memo
-          ? {
-              fields: Object.fromEntries(
-                Object.entries(sw.memo).map(([k, v]) => [
-                  k,
-                  { data: Buffer.from(JSON.stringify(v), 'utf-8') },
-                ])
-              ),
-            }
-          : undefined,
-        searchAttributes: sw.searchAttributes
-          ? {
-              indexedFields: Object.fromEntries(
-                Object.entries(sw.searchAttributes).map(([k, v]) => [
-                  k,
-                  { data: Buffer.from(JSON.stringify(v), 'utf-8') },
-                ])
-              ),
-            }
-          : undefined,
-      },
-    },
+    action: { startWorkflow },
 
     // Schedule policies
     policies: {
